requirements.txt
PyGithub
yaml
requests